LPD-79446 Create a component for date filtering
